chore(server): rename logging service
