+++ server/src/Graphql/Plugins/ApolloServerPluginLogging.ts
@@ -1,11 +1,11 @@
 import type { ApolloServerPlugin } from "@apollo/server";
-import { LoggerService } from "@Services/Logger/LoggerService";
-import { GraphLoggerOpts } from "@Services/Logger/GraphLoggerOps";
+import { LoggerService } from "@Services/LoggingService/Service";
+import { GraphLoggerOptions } from "@Services/LoggingService/GraphLoggerOptions";
 
 export const ApolloServerPlugingLogging: ApolloServerPlugin = {
   async requestDidStart({ request }) {
     if (request.query?.includes("IntrospectionQuery")) return;
-    const logger = LoggerService.getLogger(GraphLoggerOpts);
+    const logger = LoggerService.getLogger(GraphLoggerOptions);
     logger.info(`Incoming request: ${request.query}`);
   },
 };
