+++ server/src/Infrastructure/Services/LoggingService/Service.ts
@@ -1,10 +1,10 @@
 import winston from "winston";
-import type { ILoggerOptions } from "./ILoggerOptions";
+import type { LoggerOptions } from "./LoggerOptions";
 
 export class LoggerService {
   private static console = new winston.transports.Console();
 
-  public static getLogger(options: ILoggerOptions): winston.Logger {
+  public static getLogger(options: LoggerOptions): winston.Logger {
     return winston.createLogger({
       level: "info",
       format: options.format,
